examples/Python/ChatApp/requirements.txt
Flask==2.3.3
azure-identity==1.16.1
azure-appconfiguration-provider==1.3.0
openai==1.6.1